+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 * Regenerate SDK against tf2pulumi 0.6.0
+* Renamed Compute `MangedSslCertificate` to `ManagedSslCertificate`
 
 ---